src/web/requirements.txt
Flask==2.2.2
requests==2.28.2
python-dotenv==0.21.0
pyunsplash==1.0.0rc2